ntp: import RedHat patch to fix CVE-2014-9293
If no auth key is set in the configuration file, ntpd would generate
a random key on the fly. There were two problems with this: 1) the
generated key was 31 bits in size, and 2) it used the (now weak)
ntp_random() function, which was seeded with a 32 bit value and can
only provide 32 bits of entropy. This was sufficient back in the late
1990s when this code was written. Not today.

This vulnerability was discovered in ntp-4.2.6 by Neel Mehta of the
Google Security Team.

Bug #418 http://bugzilla.vyos.net/show_bug.cgi?id=418