ntp: import RedHat patch to fix CVE-2014-9295
When Autokey Authentication is enabled (i.e. the ntp.conf file contains
a crypto pw ... directive) a remote attacker can send a carefully
crafted packet that can overflow a stack buffer and potentially allow
malicious code to be executed with the privilege level of the ntpd
process.

A remote attacker can send a carefully crafted packet that can overflow
a stack buffer and potentially allow malicious code to be executed with
the privilege level of the ntpd process.

This vulnerability was discovered by Stephen Roettger of the Google
Security Team.

Bug #420 http://bugzilla.vyos.net/show_bug.cgi?id=420